Updated datatype in Addition.java

+++ Hello/src/Addition.java
@@ -4,8 +4,8 @@ public class Addition {
 
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
-		int a=sc.nextInt();
-		int b=sc.nextInt();
+		float a=sc.nextFloat();
+		float b=sc.nextFloat();
 		System.out.println("Result: "+(a+b));
 	}
